requirements.txt
asn1crypto==0.24.0
audioread==2.1.6
certifi==2018.11.29
cffi==1.11.5
chardet==3.0.4
cryptography==2.4.2
decorator==4.3.0
idna==2.8
librosa==0.6.2
llvmlite==0.27.0
nexmo==2.3.0
pycparser==2.19
PyJWT==1.7.1
pyOpenSSL==18.0.0
PySocks==1.6.8
pytz==2018.9
requests==2.21.0
six==1.12.0
tornado==5.1.1
urllib3==1.24.1
webrtcvad==2.0.10
python-dotenv
google-cloud-storage==1.13.2
